[TM-978] Move add Data button, and stop drawing event
